Remove AudioCVT interface in favor of SDL_AudioStream
